Refactor code, removed left overs
