+++ e2e/tests/loopRequest.test.ts
@@ -1,5 +1,5 @@
 import { RmfLauncher } from '../rmf-launcher';
-import { overwriteClick } from './utils';
+import { overwriteClick, getRobotLocations } from './utils';
 
 describe('loop request', () => {
   const launcher = new RmfLauncher();
@@ -17,15 +17,7 @@ describe('loop request', () => {
       return document.querySelectorAll('[data-component=RobotItem]');
     });
 
-    let getRobotLocations = () => {
-      const allRobotItems = browser.custom$$('findAllRobots', '[data-component=RobotItem]');
-      let robotLocations = allRobotItems.map(robot => {
-        robot.click();
-        return robot.$('[data-role=position]').getHTML();
-      });
-      return robotLocations;
-    };
-    const currentRobotLocations = getRobotLocations();
+    const currentRobotLocations = getRobotLocations(browser);
 
     const backButton = $('[name="back-button"]');
     backButton.click();
