+++ src/components/commands-panel.tsx
@@ -12,6 +12,22 @@ import { LoopForm } from './loop-form';
 import { v4 as uuidv4 } from 'uuid';
 import { RobotDeliveryForm } from './delivery-form';
 
+export type TDeliveryRequest = (
+  pickupPlaceName: string,
+  pickupDispenser: string,
+  dropOffPlaceName: string,
+  dropOffDispenser: string,
+  pickupBehaviour?: RomiCore.Behavior,
+  dropOffBehavior?: RomiCore.Behavior,
+) => void;
+
+export type TLoopRequest = (
+  fleetName: string,
+  numLoops: number,
+  startLocationPoint: string,
+  endLocationPoint: string,
+) => void;
+
 /**
   * task_id: Is intended to be a pseudo-random string generated by the caller which can be used to
   * identify this task as it moves between the queues to completion (or failure).
@@ -49,14 +65,6 @@ export function requestLoop(
 * dropoff_place_name: Named waypoint where the robot drops off the item. A "dropoff_dispenser" workcell is located here.
 * dropoff_dispenser: Name of the workcell unloading item from the robot at dropoff_place_name.
 */
-export type TDeliveryRequest = (
-  pickupPlaceName: string,
-  pickupDispenser: string,
-  dropOffPlaceName: string,
-  dropOffDispenser: string,
-  pickupBehaviour?: RomiCore.Behavior,
-  dropOffBehavior?: RomiCore.Behavior,
-) => void;
 
 export function requestDelivery(
   deliveryRequestPub: RomiCore.Publisher<RomiCore.Delivery> | null,
@@ -92,7 +100,7 @@ export interface CommandsPanelProps {
 
 export default function CommandsPanel(props: CommandsPanelProps): React.ReactElement {
   const { fleets, transport } = props;
-  const classes = style();
+  const classes = useStyles();
   const allFleets = fleets.flatMap(fleet => fleet.name);
   const loopRequestPub = React.useMemo(
     () => (transport ? transport.createPublisher(RomiCore.loopRequests) : null),
@@ -159,7 +167,7 @@ export default function CommandsPanel(props: CommandsPanelProps): React.ReactEle
   );
 }
 
-export const style = makeStyles(theme => ({
+export const useStyles = makeStyles(theme => ({
   expansionSummaryContent: {
     alignItems: 'center',
     justifyContent: 'space-between',
